+++ src/trait_impls.rs
@@ -1,5 +1,6 @@
 mod arrays;
 mod cells;
+mod core_types;
 mod primitives;
 mod refs;
 mod tuples;