+++ src/trait_impls/arrays.rs
@@ -14,12 +14,12 @@ macro_rules! impl_float_eq_traits_for_array {
             type Output = [<T as $crate::FloatDiff>::Output; $n];
 
             #[inline]
-            fn abs_diff(&self, other: &Self) -> Self::Output {
+            fn abs_diff(&self, other: &Self) -> Option<Self::Output> {
                 let mut result: Self::Output = unsafe { MaybeUninit::uninit().assume_init() };
                 for i in 0..$n {
-                    result[i] = self[i].abs_diff(&other[i])
+                    result[i] = self[i].abs_diff(&other[i])?
                 }
-                result
+                Some(result)
             }
 
             #[inline]