+++ tests/unit_tests.rs
@@ -69,6 +69,7 @@ impl_test_helpers!(f64, u64, crate::F64_NAN_BITS);
 mod unit_tests {
     mod arrays;
     mod cells;
+    mod core_types;
     mod macros;
     mod primitives;
     mod refs;